+++ config_test.go
@@ -214,3 +214,152 @@ func Test_assignFieldMapping(t *testing.T) {
 		})
 	}
 }
+
+type dummyRecord map[FieldIndex]string
+
+func (r dummyRecord) Parse([]byte, Metadata) error {
+	return nil
+}
+func (r dummyRecord) ToText(buf []byte) []byte {
+	return []byte("")
+}
+func (r dummyRecord) Copy() Record {
+	return Record(r.Copy())
+}
+func (r dummyRecord) Clear() {
+	r.Clear()
+}
+func (r dummyRecord) Get(i FieldIndex) []byte {
+	v, ok := r[i]
+	if !ok {
+		return make([]byte, 0)
+	}
+	return []byte(v)
+}
+func (r dummyRecord) Set(i FieldIndex, b []byte) {
+	r[i] = string(b)
+}
+func (r dummyRecord) Meta(key string) (v interface{}, ok bool) {
+	return r, true
+}
+func (r dummyRecord) Cache() *Cache {
+	return &Cache{}
+}
+
+func Test_assignValidationMapping(t *testing.T) {
+	fieldNames := []string{"name0", "name1"}
+	fieldByName := func(n string) (FieldIndex, bool) {
+		for idx, name := range fieldNames {
+			if n == name {
+				return FieldIndex(idx), true
+			}
+		}
+
+		return 0, false
+	}
+
+	cfgValidation := ConfigValidation{"name0": "^val$", "name1": "^val$"}
+	validate := func(r Record) (bool, FieldIndex) {
+		if "val" != string(r.Get(0)) {
+			return false, 0
+		}
+		if "val" != string(r.Get(1)) {
+			return false, 1
+		}
+		return true, 0
+	}
+
+	tests := []struct {
+		name      string
+		cfg       *Config
+		comp      Components
+		wantErr   bool
+		skipCheck bool
+	}{
+		{
+			name: "only in Config",
+			cfg: &Config{
+				Validation:  cfgValidation,
+				fieldByName: fieldByName, // needed by func assignValidationMapping
+			},
+			comp: Components{},
+		},
+		{
+			name: "only in Components",
+			cfg: &Config{
+				fieldByName: fieldByName, // needed by func assignValidationMapping
+			},
+			comp: Components{
+				Validate: validate,
+			},
+		},
+		{
+			name:      "nothing set",
+			cfg:       &Config{},
+			comp:      Components{},
+			skipCheck: true, // check only that cfg.validate is not nil
+		},
+
+		// error cases
+		{
+			name: "validation set both in Config and Components",
+			cfg: &Config{
+				Validation:  cfgValidation,
+				fieldByName: fieldByName, // needed by func assignValidationMapping
+			},
+			comp: Components{
+				Validate: validate,
+			},
+			wantErr: true,
+		},
+		{
+			name: "not existing field name",
+			cfg: &Config{
+				Validation:  ConfigValidation{"badname": "^val$"},
+				fieldByName: fieldByName, // needed by func assignValidationMapping
+			},
+			comp:    Components{},
+			wantErr: true,
+		},
+		{
+			name: "validation regex not compile",
+			cfg: &Config{
+				Validation:  ConfigValidation{"name0": "*"},
+				fieldByName: fieldByName, // needed by func assignValidationMapping
+			},
+			comp:    Components{},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if err := assignValidationMapping(tt.cfg, tt.comp); (err != nil) != tt.wantErr {
+				t.Errorf("assignValidationMapping() error = %v, wantErr %v", err, tt.wantErr)
+			}
+			if tt.wantErr {
+				return
+			}
+
+			if tt.cfg.validate == nil {
+				t.Errorf("cfg.validate should always be set")
+			}
+			if tt.skipCheck {
+				return
+			}
+
+			// Now we check that validate has been set correctly.
+			rec := dummyRecord{0: "val", 1: "val"}
+			if ok, field := tt.cfg.validate(rec); !ok || field != 0 {
+				t.Errorf(`cfg.validate("%v") = %v,%v, want %v,%v`, rec, ok, field, true, 0)
+			}
+			rec = dummyRecord{0: "badval", 1: "val"}
+			if ok, field := tt.cfg.validate(rec); ok && field != 0 {
+				t.Errorf(`cfg.validate("%v") = %v,%v, want %v,%v`, rec, ok, field, false, 0)
+			}
+			rec = dummyRecord{0: "val", 1: "badval"}
+			if ok, field := tt.cfg.validate(rec); ok && field != 1 {
+				t.Errorf(`cfg.validate("%v") = %v,%v, want %v,%v`, rec, ok, field, false, 1)
+			}
+		})
+	}
+}